pyproject.toml
[project]
name = "transportation"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "aiosqlite==0.20.0",
    "email-validator==2.2.0",
    "fastapi==0.115.4",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "httpx==0.28.1",
    "openpyxl==3.1.5",
    "pydantic-settings==2.6.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.20",
    "redis==5.2.1",
    "reportlab==4.2.0",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.32.0",
]

[project.optional-dependencies]
dev = [
    "mypy==1.13.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "ruff==0.7.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["website/static"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
pretty = true
show_error_codes = true
exclude = "^(website/|sikar_cargo_transport/)"

[[tool.mypy.overrides]]
module = ["tests.*", "alembic.*"]
allow_untyped_defs = true
allow_incomplete_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false

[[tool.mypy.overrides]]
module = ["jose", "jose.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "openpyxl",
    "openpyxl.*",
    "reportlab",
    "reportlab.*",
    "svglib",
    "svglib.*",
    "cairosvg",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q --cov=src --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["tests/*", "alembic/*"]

[tool.coverage.report]
skip_empty = true
show_missing = true
fail_under = 70
